packages

+++ src-bill/lists/GeoPrefixList.h
@@ -42,7 +42,7 @@ public:
     GeoPrefix * find(long long int prefix, stringstream *trace = nullptr) {
         char tmpPrefix[20];
         sprintf(tmpPrefix, "%lld", prefix);
-        int len = strlen(tmpPrefix);
+        size_t len = strlen(tmpPrefix);
         while (len > 0) {
             tmpPrefix[len] = 0;
             auto result = _find(tmpPrefix);
