+++ src-bill/lists/MobPrefixList.h
@@ -41,7 +41,7 @@ public:
     MobPrefix * find(long long int prefix, stringstream *trace = nullptr) {
         char tmpPrefix[20];
         sprintf(tmpPrefix, "%lld", prefix);
-        int len = strlen(tmpPrefix);
+        size_t len = strlen(tmpPrefix);
         while (len > 0) {
             tmpPrefix[len] = 0;
             auto result = _find(tmpPrefix);
