+++ src-bill/lists/NetworkPrefixList.h
@@ -75,7 +75,7 @@ public:
         char tmpPrefix[20];
         sprintf(tmpPrefix, "%lld", numberPrefix);
 
-        int len = strlen(tmpPrefix);
+        size_t len = strlen(tmpPrefix);
         while (len > 0) {
             tmpPrefix[len] = 0;
