+++ src-bill/lists/PrefixlistPrefixList.h
@@ -55,7 +55,7 @@ public:
     PrefixlistPrefix * find(int prefixlist_id, char * prefix) {
         char tmpPrefix[20];
         strcpy(tmpPrefix, prefix);
-        int len = strlen(tmpPrefix);
+        size_t len = strlen(tmpPrefix);
         while (len > 0) {
             tmpPrefix[len] = 0;
             auto result = _find(prefixlist_id, tmpPrefix);
